horizon_web_ui/test-requirements.txt
-e git+https://github.com/openstack/horizon.git#egg=horizon
django-nose
nose-exclude
mock
python-openstackclient
mox
parsedatetime
coverage>=3.6
